Extract common methods
