+++ nirdizati/encoders/tests/test_boolean.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 
 from nirdizati.encoders import boolean
-from nirdizati.encoders.tests.helper import data_frame
+from nirdizati.encoders.tests.setup import data_frame
 
 
 class TestBoolean(TestCase):
