+++ nirdizati/encoders/tests/test_complex.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 
 from nirdizati.encoders import complex
-from nirdizati.encoders.tests.helper import data_frame, get_encoder
+from nirdizati.encoders.tests.setup import data_frame, get_encoder
 
 
 class TestComplex(TestCase):
