+++ nirdizati/encoders/tests/test_index_latest.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 
 from nirdizati.encoders import index_latest_payload
-from nirdizati.encoders.tests.helper import data_frame, get_encoder
+from nirdizati.encoders.tests.setup import data_frame, get_encoder
 
 
 class TestIndexLatestPayload(TestCase):
